+++ src/OpenLoco/CMakeLists.txt
@@ -246,6 +246,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/OpenLoco.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/Paint.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintAirport.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintBridge.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintBuilding.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintDocks.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintEffectEntity.cpp"
@@ -607,6 +608,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/OpenLoco.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/Paint.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintAirport.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintBridge.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintBuilding.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintDocks.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintEffectEntity.h"
