+++ src/OpenLoco/src/Paint/PaintBridge.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace OpenLoco::Paint
+{
+    struct PaintSession;
+
+    bool sub_42AC9C(PaintSession& session);
+}
